0112leesy-11-12-B008-rename apis to views
